Function wake_on_lan_hook::server::run

pub fn run(
    log: Logger, 
    desired_mac_address: MacAddress, 
    cmd: Vec<String>
) -> Result<(), Error>

Run the wake-on-lan-hook server.

This will start listening on UDP ports 0, 7, and 9 for wake-on-LAN "magic packets" and run the given command whenever a packet for the desired MAC address is detected.

Wake-on-LAN packets for other MAC addresses will be ignored but logged.

See the magic_packet() parser for details about what constitutes a magic packet.